+++ workbench/database/migrations/create_inbound_emails_table.php
@@ -26,7 +26,10 @@ return new class extends Migration
             $table->timestamp('received_at')->nullable();
 
             $table->string('organization_alias')->nullable();
-            $table->unsignedBigInteger('tenant_id')->nullable();
+
+            if (config('inbound-email.multi_tenant_enabled')) {
+                $table->foreignId('tenant_id')->nullable()->index();
+            }
 
             $table->string('status');
             $table->text('error')->nullable();
@@ -40,7 +43,6 @@ return new class extends Migration
             $table->index('status');
             $table->index('message_id');
             $table->index('organization_alias');
-            $table->index('tenant_id');
         });
     }